requirements.txt
# Web Scraping
selenium>=4.15.0
webdriver-manager>=4.0.1
beautifulsoup4>=4.12.0
requests>=2.31.0
lxml>=4.9.0

# AI — Claude API (primary) + Ollama (local fallback)
anthropic>=0.96.0
ollama>=0.6.0

# Email report
matplotlib>=3.7.0

# Twitter API
requests-oauthlib>=1.3.0

# Utilities
python-dotenv>=1.0.0
pandas>=2.1.0

# PDF Processing
PyPDF2>=3.0.0
